+++ co-log-core/CHANGELOG.md
@@ -3,6 +3,11 @@
 `co-log-core` uses [PVP Versioning][1].
 The change log is available [on GitHub][2].
 
+## Unreleased
+
+* [#122](https://github.com/kowainik/co-log/issues/122):
+  Add the `separate` combinator.
+
 ## 0.2.0.0 — May 5, 2019
 
 * [#85](https://github.com/kowainik/co-log/issues/85):